Drop poise as a dependency
While it provided some useful features, it also provided/encouraged patterns that seem to muddy the intended use of the Discord API, so I found it somewhat lacking.

This is a mostly working replacement; it also updates some commands to use components with better control over their selections than parameters have. There are a couple that still need fixed to actually work with the Discord API, and some that need revised to handle more than 25 values (which will happen).